# src/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CC REST API",
    "description": ""
  },
  "host": "",
  "basePath": "/api",
  "tags": [
    {
      "name": "Users",
      "description": "User administration"
    },
    {
      "name": "Root",
      "description": "Fallback URL"
    },
    {
      "name": "Certificates",
      "description": "Certificates post/retrieve"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Root"
        ],
        "summary": "Shows HTML for initial page",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register a user to the system",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "userName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "passwordConfirmation": {
                  "example": "any"
                },
                "profilePic": {
                  "example": "any"
                }
              }
            },
            "description": "Adding new user."
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Logs in a registered user",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            },
            "description": "Adding new user."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get loged users profile",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "A JWT bearer",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained."
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Lists all users in database",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/certificates/upload": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "summary": "Uploads a certificate",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "certificate": {
                  "example": "any"
                }
              }
            },
            "description": "Uploads a new certificate."
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/certificates/list": {
      "get": {
        "tags": [
          "Certificates"
        ],
        "summary": "Get logged in user's certificates",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "A JWT bearer",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Certificates listed."
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                }
              }
            },
            "firstName": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "lastName": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "username": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "email": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "password": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "profilePic": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "binary"
                }
              }
            }
          }
        }
      }
    },
    "Certificate": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                }
              }
            },
            "certificate": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "binary"
                }
              }
            },
            "userId": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "examples": {
      "User": {
        "value": {
          "id": 1,
          "firstName": "foo",
          "lastName": "bar",
          "username": "foobar",
          "email": "foo@bar.com",
          "password": "foobarium",
          "profilePic": "...base64 encoded image..."
        }
      },
      "Certificate": {
        "value": {
          "id": 1,
          "certificate": "...base64 encoded certificate...",
          "userId": 1
        }
      }
    }
  }
}